+++ user_docs/en/changes.md
@@ -29,9 +29,8 @@ Unicode CLDR has also been updated.
   * This can be disabled in the "Add-on Store" category of settings.
   * NVDA checks daily for add-on updates.
   * Only updates within the same channel will be checked (e.g. installed beta add-ons will only notify for updates in the beta channel).
-* By default, after NVDA startup, you will be notified if any new add-on is available. (#16681, @nvdaes)
-  * This can be disabled in the "Add-on Store" category of settings.
-  * You can also check for new add-ons from the New available add-ons of the store.
+* You can check for new add-ons from the New available add-ons tab of the store. (#16681, @nvdaes)
+  * From the Add-on store settings panel, new options are available to determine if the presence of new add-ons should be notified at startup, and the period of time to reset new add-ons.
 
 ### Changes